*lets split them
