+++ src/lib.rs
@@ -17,4 +17,5 @@ pub mod mpegts;
 pub mod script_wasm;
 pub mod usb_gadget;
 pub mod usb_stream;
+pub mod vendor_ext;
 pub mod web;
